Removing Retrofit
