+++ http/build.gradle
@@ -1,7 +1,7 @@
 apply plugin: 'com.android.library'
 apply plugin: 'com.github.dcendents.android-maven'
-group = 'com.github.amitclickapps'
-version = '1.0.7'
+group = 'com.github.amitsahni'
+version = '1.0.0'
 
 android {
     compileSdkVersion 27
@@ -24,9 +24,8 @@ android {
 
 dependencies {
     compile 'com.android.support:support-annotations:25.3.1'
-    /*  Web Calling Retrofit + Okhttp */
-    compile 'com.squareup.retrofit2:retrofit:2.3.0'
-    compile 'com.squareup.retrofit2:converter-gson:2.3.0'
+    /* Okhttp */
+    compile 'com.squareup.okhttp3:okhttp:3.9.1'
     compile 'com.squareup.okhttp3:logging-interceptor:3.9.1'
 //    // Reactive Android and Java
 //    compile 'io.reactivex.rxjava2:rxandroid:2.0.1'
