+++ http/src/main/java/webconnect/com/webconnect/WebConnect.java
@@ -1,6 +1,5 @@
 package webconnect.com.webconnect;
 
-import android.app.Activity;
 import android.content.Context;
 import android.support.annotation.NonNull;
 
@@ -29,18 +28,6 @@ public class WebConnect {
             }
         }
         return sWebConnect;
-    }
-
-    /**
-     * With builder.
-     *
-     * @param context the context
-     * @param url     the url
-     * @return the builder
-     */
-    public static Builder with(@NonNull Activity context, @NonNull String url) {
-        return new Builder(context, url);
-
     }
 
     /**
